+++ src/lightly_train/_task_models/dinov2_eomt_instance_segmentation/train_model.py
@@ -22,6 +22,7 @@ from torch.optim.optimizer import Optimizer
 from lightly_train import _torch_helpers
 from lightly_train._configs.validate import no_auto
 from lightly_train._data.instance_segmentation_dataset import (
+    COCOInstanceSegmentationDataArgs,
     YOLOInstanceSegmentationDataArgs,
 )
 from lightly_train._data.task_data_args import TaskDataArgs
@@ -166,7 +167,7 @@ class DINOv2EoMTInstanceSegmentationTrain(TrainModel):
         *,
         model_name: str,
         model_args: DINOv2EoMTInstanceSegmentationTrainArgs,
-        data_args: YOLOInstanceSegmentationDataArgs,
+        data_args: COCOInstanceSegmentationDataArgs | YOLOInstanceSegmentationDataArgs,
         train_transform_args: DINOv2EoMTInstanceSegmentationTrainTransformArgs,
         val_transform_args: DINOv2EoMTInstanceSegmentationValTransformArgs,
         load_weights: bool,